requirements.txt
colorama==0.4.4
glfw==2.5.3
Nuitka @ https://github.com/Nuitka/Nuitka/archive/factory.zip
numpy==1.22.3
ordered-set==4.1.0
pygame==2.1.2
PyOpenGL==3.1.6
PyOpenGL-accelerate==3.1.6
Rafflesia==0.0.9.8
tqdm==4.64.0
zstandard==0.17.0
[0m